# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: preferences
    description: User taste & wellness preferences
  - name: wellness
    description: Wellness dashboard, wearables, and daily activity sync
  - name: recipes
    description: RD/chef-authored recipe library
  - name: challenges
    description: Cohort challenges with shared progress feed
  - name: reviews
    description: Per-dish customer reviews and moderation
paths:
  /recipes:
    get:
      operationId: listRecipes
      tags: [recipes]
      summary: List published recipes with optional filters
      parameters:
        - in: query
          name: goal
          schema: { type: string }
        - in: query
          name: diet
          schema: { type: string }
        - in: query
          name: maxTime
          schema: { type: integer, minimum: 1 }
        - in: query
          name: q
          schema: { type: string }
      responses:
        "200":
          description: Recipes
          content:
            application/json:
              schema:
                type: object
                required: [recipes]
                properties:
                  recipes:
                    type: array
                    items: { $ref: "#/components/schemas/Recipe" }
  /recipes/{slug}:
    get:
      operationId: getRecipe
      tags: [recipes]
      summary: Get a recipe by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Recipe
          content:
            application/json:
              schema:
                type: object
                required: [recipe]
                properties:
                  recipe: { $ref: "#/components/schemas/Recipe" }
        "404":
          description: Not found
  /challenges:
    get:
      operationId: listChallenges
      tags: [challenges]
      summary: List cohort challenges
      responses:
        "200":
          description: Challenges
          content:
            application/json:
              schema:
                type: object
                required: [challenges]
                properties:
                  challenges:
                    type: array
                    items: { $ref: "#/components/schemas/Challenge" }
  /challenges/{slug}:
    get:
      operationId: getChallenge
      tags: [challenges]
      summary: Get a challenge with cohort feed
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Challenge detail
          content:
            application/json:
              schema:
                type: object
                required: [challenge, joined, posts]
                properties:
                  challenge: { $ref: "#/components/schemas/Challenge" }
                  joined: { type: boolean }
                  posts:
                    type: array
                    items: { $ref: "#/components/schemas/ChallengePost" }
        "404":
          description: Not found
  /challenges/{slug}/join:
    post:
      operationId: joinChallenge
      tags: [challenges]
      summary: Join a challenge (auth required)
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
  /challenges/{slug}/leave:
    post:
      operationId: leaveChallenge
      tags: [challenges]
      summary: Leave a challenge (auth required)
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
  /challenges/{slug}/posts:
    post:
      operationId: postToChallenge
      tags: [challenges]
      summary: Post to the cohort feed (members only)
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengePostInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [post]
                properties:
                  post: { $ref: "#/components/schemas/ChallengePost" }
        "401":
          description: Unauthorized
        "403":
          description: Not a member
  /dish-reviews:
    post:
      operationId: createDishReview
      tags: [reviews]
      summary: Submit a review for a dish (auth required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DishReviewInput"
      responses:
        "200":
          description: Created
        "401":
          description: Unauthorized
  /dish-reviews/{slug}:
    get:
      operationId: listDishReviews
      tags: [reviews]
      summary: Public reviews and AI summary for a dish
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Reviews
  /dish-reviews-mod:
    get:
      operationId: listReviewsForModeration
      tags: [reviews]
      summary: Catalog-scoped — list recent reviews including hidden
      responses:
        "200":
          description: Reviews
        "403":
          description: Catalog scope required
  /dish-reviews/{id}/hide:
    post:
      operationId: hideDishReview
      tags: [reviews]
      summary: Catalog-scoped — hide a review
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated
        "403":
          description: Catalog scope required
        "404":
          description: Not found
  /dish-reviews/{id}/unhide:
    post:
      operationId: unhideDishReview
      tags: [reviews]
      summary: Catalog-scoped — unhide a review
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated
        "403":
          description: Catalog scope required
        "404":
          description: Not found
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /preferences:
    get:
      operationId: getPreferences
      tags: [preferences]
      summary: Get the signed-in user's preferences
      responses:
        "200":
          description: Current preferences (null when unset)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreferencesEnvelope"
        "401":
          description: Unauthorized
    put:
      operationId: putPreferences
      tags: [preferences]
      summary: Upsert preferences (partial)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreferencesPatch"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreferencesEnvelopeRequired"
        "400":
          description: Invalid payload
        "401":
          description: Unauthorized
    patch:
      operationId: patchPreferences
      tags: [preferences]
      summary: Partial update of preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreferencesPatch"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreferencesEnvelopeRequired"
        "400":
          description: Invalid payload
        "401":
          description: Unauthorized
  /wellness/today:
    get:
      operationId: getWellnessToday
      tags: [wellness]
      summary: Get today's wellness summary (targets, totals, wearables, streaks)
      responses:
        "200":
          description: Today's wellness snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WellnessTodayEnvelope"
        "401":
          description: Unauthorized
  /wellness/wearable/connect:
    post:
      operationId: connectWearable
      tags: [wellness]
      summary: Mark a wearable provider as connected for the signed-in user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WearableConnectInput"
      responses:
        "200":
          description: Connected wearable link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WearableLinkEnvelope"
        "400":
          description: Invalid payload
        "401":
          description: Unauthorized
  /wellness/wearable/disconnect:
    post:
      operationId: disconnectWearable
      tags: [wellness]
      summary: Mark a wearable provider as disconnected for the signed-in user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WearableConnectInput"
      responses:
        "200":
          description: Disconnect ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkEnvelope"
        "400":
          description: Invalid payload
        "401":
          description: Unauthorized
  /wellness/wearable/sync:
    post:
      operationId: syncWearable
      tags: [wellness]
      summary: Push today's activity (steps + active kcal) from a wearable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WearableSyncInput"
      responses:
        "200":
          description: Updated wearable link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WearableLinkEnvelope"
        "400":
          description: Invalid payload
        "401":
          description: Unauthorized
        "409":
          description: Wearable not connected
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    DietaryStyle:
      type: string
      enum: [omnivore, vegetarian, vegan, pescatarian, keto]
    SpiceLevel:
      type: string
      enum: [none, mild, medium, hot]
    ActivityLevel:
      type: string
      enum: [sedentary, light, moderate, active, very_active]
    WellnessGoal:
      type: string
      enum: [lose_weight, maintain, gain_muscle, general_wellness]
    UserPreferences:
      type: object
      required:
        - userId
        - allergens
        - dislikedIngredients
        - cuisines
        - spiceLevel
        - dietaryStyle
        - goal
        - activityLevel
        - createdAt
        - updatedAt
      properties:
        userId:
          type: string
        allergens:
          type: array
          items: { type: string }
        dislikedIngredients:
          type: array
          items: { type: string }
        cuisines:
          type: array
          items: { type: string }
        spiceLevel:
          $ref: "#/components/schemas/SpiceLevel"
        dietaryStyle:
          $ref: "#/components/schemas/DietaryStyle"
        goal:
          $ref: "#/components/schemas/WellnessGoal"
        activityLevel:
          $ref: "#/components/schemas/ActivityLevel"
        calorieTarget:
          type: integer
          nullable: true
          minimum: 800
          maximum: 6000
        proteinTargetGrams:
          type: integer
          nullable: true
          minimum: 20
          maximum: 400
        carbsTargetGrams:
          type: integer
          nullable: true
          minimum: 0
          maximum: 800
        fatTargetGrams:
          type: integer
          nullable: true
          minimum: 0
          maximum: 300
        quizCompletedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PreferencesPatch:
      type: object
      properties:
        allergens:
          type: array
          items: { type: string, minLength: 1, maxLength: 64 }
          maxItems: 50
        dislikedIngredients:
          type: array
          items: { type: string, minLength: 1, maxLength: 64 }
          maxItems: 50
        cuisines:
          type: array
          items: { type: string, minLength: 1, maxLength: 64 }
          maxItems: 50
        spiceLevel:
          $ref: "#/components/schemas/SpiceLevel"
        dietaryStyle:
          $ref: "#/components/schemas/DietaryStyle"
        goal:
          $ref: "#/components/schemas/WellnessGoal"
        activityLevel:
          $ref: "#/components/schemas/ActivityLevel"
        calorieTarget:
          type: integer
          nullable: true
          minimum: 800
          maximum: 6000
        proteinTargetGrams:
          type: integer
          nullable: true
          minimum: 20
          maximum: 400
        carbsTargetGrams:
          type: integer
          nullable: true
          minimum: 0
          maximum: 800
        fatTargetGrams:
          type: integer
          nullable: true
          minimum: 0
          maximum: 300
        markQuizComplete:
          type: boolean
    PreferencesEnvelope:
      type: object
      required: [preferences]
      properties:
        preferences:
          oneOf:
            - $ref: "#/components/schemas/UserPreferences"
            - type: "null"
    PreferencesEnvelopeRequired:
      type: object
      required: [preferences]
      properties:
        preferences:
          $ref: "#/components/schemas/UserPreferences"
    WearableProvider:
      type: string
      enum: [apple_health, google_fit]
    WearableConnectInput:
      type: object
      required: [provider]
      properties:
        provider:
          $ref: "#/components/schemas/WearableProvider"
    WearableSyncInput:
      type: object
      required: [provider, activityKcal]
      properties:
        provider:
          $ref: "#/components/schemas/WearableProvider"
        activityKcal:
          type: integer
          minimum: 0
          maximum: 3000
        steps:
          type: integer
          minimum: 0
          maximum: 60000
    WearableLink:
      type: object
      required: [id, userId, provider, connected]
      properties:
        id:
          type: string
        userId:
          type: string
        provider:
          $ref: "#/components/schemas/WearableProvider"
        connected:
          type: boolean
        lastSyncedAt:
          type: string
          format: date-time
          nullable: true
        lastActivityKcal:
          type: integer
          nullable: true
        lastSteps:
          type: integer
          nullable: true
    WearableLinkEnvelope:
      type: object
      required: [link]
      properties:
        link:
          $ref: "#/components/schemas/WearableLink"
    OkEnvelope:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    WellnessDailyTotals:
      type: object
      required: [date, calories, proteinGrams]
      properties:
        date:
          type: string
        calories:
          type: number
        proteinGrams:
          type: number
        carbsGrams:
          type: number
        fatGrams:
          type: number
        fiberGrams:
          type: number
        waterMl:
          type: number
        vegServings:
          type: number
    WellnessTargets:
      type: object
      required: [calorieTarget, proteinTargetGrams, effectiveCalorieTarget, activityKcal]
      properties:
        calorieTarget:
          type: integer
        proteinTargetGrams:
          type: integer
        fiberTargetGrams:
          type: integer
        waterTargetMl:
          type: integer
        vegTargetServings:
          type: integer
        effectiveCalorieTarget:
          type: integer
        activityKcal:
          type: integer
    WellnessTodayEnvelope:
      type: object
      required: [date, targets, totals, wearables]
      properties:
        date:
          type: string
        targets:
          $ref: "#/components/schemas/WellnessTargets"
        totals:
          $ref: "#/components/schemas/WellnessDailyTotals"
        wearables:
          type: array
          items:
            $ref: "#/components/schemas/WearableLink"
    Recipe:
      type: object
      required: [id, slug, title, summary, goalTags, dietTags, totalTimeMinutes, status]
      properties:
        id: { type: integer }
        slug: { type: string }
        title: { type: string }
        summary: { type: string }
        heroImageUrl: { type: string, nullable: true }
        authorName: { type: string, nullable: true }
        authorRole: { type: string, nullable: true }
        goalTags:
          type: array
          items: { type: string }
        dietTags:
          type: array
          items: { type: string }
        totalTimeMinutes: { type: integer }
        servings: { type: integer, nullable: true }
        ingredients:
          type: array
          items:
            type: object
            required: [text]
            properties:
              text: { type: string }
        steps:
          type: array
          items: { type: string }
        nutrition:
          type: object
          nullable: true
          properties:
            kcal: { type: number }
            proteinG: { type: number }
            carbsG: { type: number }
            fatG: { type: number }
            fiberG: { type: number }
        status: { type: string }
    Challenge:
      type: object
      required: [id, slug, title, summary, durationDays, status]
      properties:
        id: { type: integer }
        slug: { type: string }
        title: { type: string }
        summary: { type: string }
        heroImageUrl: { type: string, nullable: true }
        durationDays: { type: integer }
        startDate: { type: string, nullable: true }
        endDate: { type: string, nullable: true }
        bundleSlug: { type: string, nullable: true }
        rdName: { type: string, nullable: true }
        rdCheckinCadence: { type: string, nullable: true }
        memberCount: { type: integer }
        status: { type: string }
    ChallengePost:
      type: object
      required: [id, challengeId, body, createdAt]
      properties:
        id: { type: integer }
        challengeId: { type: integer }
        userId: { type: string, nullable: true }
        authorName: { type: string, nullable: true }
        body: { type: string }
        kind: { type: string, nullable: true }
        createdAt: { type: string }
    ChallengePostInput:
      type: object
      required: [body]
      properties:
        body: { type: string, minLength: 1, maxLength: 1000 }
    DishReviewInput:
      type: object
      required: [slug, rating]
      properties:
        slug: { type: string }
        rating: { type: integer, minimum: 1, maximum: 5 }
        body: { type: string, maxLength: 2000 }
        photoUrl:
          type: string
          format: uri
          maxLength: 1024
          nullable: true