# lib/api-zod/src/generated/types/challenge.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Challenge {
  id: number;
  slug: string;
  title: string;
  summary: string;
  heroImageUrl?: string | null;
  durationDays: number;
  startDate?: string | null;
  endDate?: string | null;
  bundleSlug?: string | null;
  rdName?: string | null;
  rdCheckinCadence?: string | null;
  memberCount?: number;
  status: string;
}